Fix SVG Indicator text (gfx.Text) dimensions in Firefox.
Implements a workaround for https://bugzilla.mozilla.org/show_bug.cgi?id=1109860. When the text of an SVG indicator node is modified for bidi support, Firefox miscalculates the dimensions. This workaround clones the indicator node, reapplies the same text with the bidi characters stripped out, and returns the correct `getBBox` dimensions.